==> manipulation/legacy/1-scale-assembly.R <==
# The purpose of this script is to create a data object (dto) which will hold all data and metadata.
# Run the lines below to stitch a basic html output.
# knitr::stitch_rmd(
#   script="./manipulation/0-ellis-island.R",
#   output="./manipulation/stitched-output/0-ellis-island.md"
# )
# The above lines are executed only when the file is run in RStudio, !! NOT when an Rmd/Rnw file calls it !!

############################
##  Land on Ellis Island  ##
############################

rm(list=ls(all=TRUE))  #Clear the variables from previous runs.
cat("\f") # clear console

# ---- load-sources ------------------------------------------------
# Call `base::source()` on any repo file that defines functions needed below.  
# Ideally, no real operations are performed in these sourced scripts. 
source("./scripts/functions-common.R") # used in multiple reports

# ---- load-packages ----------------------------------------------
# Attach packages so their functions don't need to be qualified when used
# See more : http://r-pkgs.had.co.nz/namespace.html#search-path
library(magrittr) # Pipes
library(ggplot2) # Graphs
# Functions of these packages will need to be qualified when used
# See more: http://r-pkgs.had.co.nz/namespace.html#search-path
requireNamespace("tidyr") #  data manipulation
requireNamespace("dplyr") # f() names conflict with other packages (esp. base, stats, and plyr).
requireNamespace("testit") # for asserting conditions meet expected patterns.

# ---- declare-globals ----------------------------------------------
path_renaming_rules <- "./data-phi-free/raw/renaming-rules/renaming-rules.xlsx"

# ---- load-data ------------------------------------------------
ds_rand <- readRDS("./data-unshared/derived/rndhrs_p.rds")
ds_2004 <- readRDS("./data-unshared/derived/h04f1a.rds")
ds_2006 <- readRDS("./data-unshared/derived/h06f2b.rds")
ds_2008 <- readRDS("./data-unshared/derived/h08f2a.rds")
ds_2010 <- readRDS("./data-unshared/derived/hd10f5c.rds")
ds_2012 <- readRDS("./data-unshared/derived/h12f1a.rds")
ds_2014 <- readRDS("./data-unshared/derived/h14e1a.rds")



# make the names lowercase. ??? This may be undesirable. New names only?
for(i in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
    # create a string to be passed as command to the eval() function
    # i <- 2004
    cstring <- paste0("colnames(ds_",i,") <- tolower(colnames(ds_",i,"))")
    eval(parse(text=cstring)) # evaluates the content of the command string
} 

# load meta-data
(path_meta <- list.files("./data-phi-free/raw/meta/RAND-2017-02-08", full.names = T, pattern=".csv$"))
ls_meta <- list()
for(i in seq_along(path_meta)){
  section_name <- sub(".csv$","", basename(path_meta[i]) )
  ls_meta[[section_name]] <- readr::read_csv(path_meta[i]) %>% 
    dplyr::mutate(
      year = as.numeric(year),
      reversed = as.numeric(reversed)
    )
  attr(ls_meta[[section_name]],"spec") <- NULL
}
ls_meta %>%  names()

# ----- dummy -----------------------
# names_labels(ds04)
# dim(ds04)


# ---- define-utility-functions -------------------
# these functions are used in all scale computations
# custom functions are placed in corresponding chunks

# renames and subsets the original data frame
subset_rename <- function(d,renaming_rules,year_){ 
  # broswer()
  items <- renaming_rules %>% 
    dplyr::filter(year==year_) %>% 
    dplyr::select(old_name,new_name) %>% 
    as.data.frame()
  # get the list of old and new names from the spreadsheet
  (old_names <- c("hhidpn",items[,"old_name"]))
  (new_names <- c("hhidpn",items[,"new_name"]))
  dnew <- d %>% 
    # dplyr::filter(hhidpn == 3010) %>% # filter a specific individual
    dplyr::select_(.dots = old_names) 
  colnames(dnew) <- new_names 
  return(dnew) 
}
colnames(ds_2012)
# standard computation of scale scores
compute_scale_score <- function(d){
  # d <- ds_long
  (col_names <- setdiff(names(d),c("year","hhidpn")))
  d[,'sum'] <-  apply(d[col_names],1,sum, na.rm = FALSE)
  d[,'mean'] <-  apply(d[col_names],1,mean, na.rm = FALSE)
  return(d)
}
# Usage:
# ds_long <- ds_long %>% compute_scale_score()


# reverse the coding on selected items
# define function to reverse code a specified variable
reverse_coding <- function(d, variables){
  # d <- ds_lone
  for(v in variables){
    # v = "loneliness_1"
    (p <- unique(d[,v]) %>% as.numeric())
    (p <- p[!is.na(p)])
    (original_values <- sort(p,decreasing = F))
    (reversed_values <- sort(p,decreasing = T))
    d[,v] <- plyr::mapvalues(d[,v], from=original_values, to=reversed_values) 
  }
  return(d)
}
# Usage:
# ds_lone <- ds_lone %>% 
  # reverse_coding(reverse_these)

# ---- create-dto ---------------------
dto <- list()

# ------ RAND ------------------------
#read in the renaming rules for this specific variables
rename_rand   <-  ls_meta[["rand"]]

# now select the rand ds 
ls_temp <- list()
subset_rename_rand <- function(year_){ 
  # broswer()
  items <- rename_rand %>% 
    dplyr::filter(year==year_) %>% 
    dplyr::select(old_name,new_name) %>% 
    as.data.frame()
  # get the list of old and new names from the spreadsheet
  (old_names <- c("hhidpn",items[,"old_name"]))
  (new_names <- c("hhidpn",items[,"new_name"]))
  dnew <- ds_rand %>% 
    # dplyr::filter(hhidpn == 3010) %>% # filter a specific individual
    dplyr::select_(.dots = old_names) 
  colnames(dnew) <- new_names 
  dnew[,"year"]<- year_
  return(dnew) 
}

ls_temp[[paste(2004)]] <- subset_rename_rand(2004)
ls_temp[[paste(2006)]] <- subset_rename_rand(2006)
ls_temp[[paste(2008)]] <- subset_rename_rand(2008)
ls_temp[[paste(2010)]] <- subset_rename_rand(2010)
ls_temp[[paste(2012)]] <- subset_rename_rand(2012)
ls_temp[[paste(2014)]] <- subset_rename_rand(2014)

# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)

ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

dto[["rand"]] <- ds_long

# ----- demographics ------------------
# path_input_map <- "./data-shared/raw/mhsu-service-types/mhsu-service-type-mapping-2016-09-02.csv"
#read in the renaming rules for this specific variables
rename_demographics   <-  ls_meta[["demographics"]]
str(rename_demographics)
# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){
  # create a string to be passed as command to the eval() function
  # year <- 2006
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_demographics,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}

# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

ds_long %>% dplyr::filter(hhidpn== "3020")

dto[["demographics"]] <- ds_long


# ----- loneliness -------------
# path_input_map <- "./data-shared/raw/mhsu-service-types/mhsu-service-type-mapping-2016-09-02.csv"
#read in the renaming rules for this specific variables
rename_loneliness   <-  ls_meta[["loneliness"]]
# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  # year <- 2006
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_loneliness,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)

ds_long %>% dplyr::filter(hhidpn==10001010)

# create a vector with names of items to be reverse scored
rename_meta <-rename_loneliness 
# use meta data to provide the rules
reverse_these <- unique( rename_meta[rename_meta$reversed==TRUE,"new_name"] ) %>% as.data.frame()
reverse_these <- reverse_these[!is.na(reverse_these)]
ds_long <- ds_long %>% 
  reverse_coding(reverse_these)
 
# d <- ds_long %>% dplyr::filter(hhidpn==10001010)
compute_loneliness_scale_score <- function(d){
  # d <- ds_long %>% dplyr::filter(hhidpn %in% c(3010,10281010))
  (col_names_11 <- setdiff(names(d),c("year","hhidpn")))
  (col_names_3 <- col_names_11[1:3])
  d[,"lonelysum_11"] <- apply(d[col_names_11],1,sum, na.rm = TRUE)
  d[,"lonelysum_3"] <- apply(d[col_names_3],1,sum, na.rm = TRUE)
  d[,"score_loneliness_3"] <- apply(d[col_names_3],1,mean, na.rm = TRUE)
  d$missing_count <- apply(d[col_names_11], 1, function(z) sum(is.na(z)))
  d <- d %>% 
    dplyr::mutate( 
      score_loneliness_11 = ifelse(missing_count<6, 
                    lonelysum_11/(11- missing_count),NA)
    )
  return(d)
}
# Usage:
ds_long <- compute_loneliness_scale_score(ds_long)
dto[["loneliness"]] <- ds_long



# ----- life_satisfaction -------------
#path_input_map <- "./data-shared/raw/mhsu-service-types/mhsu-service-type-mapping-2016-09-02.csv"
#read in the renaming rules for this specific variables
rename_life_satisfaction  <-  ls_meta[["life-satisfaction"]] %>% as.data.frame()

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  # year <- 2006
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_life_satisfaction,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

rename_meta <-rename_life_satisfaction 
reverse_these <- unique( rename_meta[rename_meta$reversed==TRUE,"new_name"] )
reverse_these <- reverse_these[!is.na(reverse_these)]
testit::assert("The scale does not contained reverse coded items",reverse_these==0L)

# d <- ds_long %>% dplyr::filter(hhidpn==10001010)
ds_long <- ds_long %>% compute_scale_score()
head(ds_long)
dto[["life_satisfaction"]] <- ds_long


# ----- social-network -------------
#read in the renaming rules for this specific variables
rename_social_network  <-  ls_meta[["social-network"]] %>% as.data.frame()
# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  # year <- 2006
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_social_network,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}

# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

networkvars <- c("snspouse", "snchild", "snfamily","snfriends")
closevars <-c("closespouse", "closechild", "closefam", "closefri")

#The social network count items do not need to be reverse coded instead it was coded as 1 if yes 
# (e.g., yes they have children) and 5 in no (e.g., no children). I wanted to recode this so that it was 1 and 0
# this allows a social network total score to be calculated.

reverse_coding_socialnetwork <- function(d, variables){
  # d <- ds_lone
  for(v in variables){
    # v = "loneliness_1"
    (p <- unique(d[,v]) %>% as.numeric())
    (p <- p[!is.na(p)])
    d[,v] <- plyr::mapvalues(d[,v], from=5, to=0) 
    d[,v] <- plyr::mapvalues(d[,v], from=7, to=NA) 
  }
  return(d)
}

ds_long <- ds_long %>% 
 reverse_coding_socialnetwork(networkvars)
#Computes two scores socialnetwork_total a count of whether or not network members exist in each of the 
# four possible categories and close_social_network a count of the total number of relationships the respondent
# considers close relationships across all relational categories.
compute_socialnetwork_scale_scores <- function(d){
  #d <- ds_long %>% dplyr::filter(hhidpn %in% c(3010,10281010))
  d[,"socialnetwork_total"] <- apply(d[networkvars],1,sum, na.rm = TRUE)
  d[,"close_social_network"] <- apply(d[closevars],1,sum, na.rm = TRUE)
  d$missing_count <- apply(d[networkvars], 1, function(z) sum(is.na(z)))
  d <- d %>% 
    dplyr::mutate( 
      socialnetwork_total = ifelse(missing_count<4, 
                                   socialnetwork_total,NA))
  d$missing_count <- apply(d[closevars], 1, function(z) sum(is.na(z)))   
   d <- d %>% 
        dplyr::mutate( 
          close_social_network = ifelse(missing_count<4, 
                                        close_social_network,NA)
    )
  return(d)
}

# d <- ds_long %>% dplyr::filter(hhidpn==10001010)
ds_long <- ds_long %>% compute_socialnetwork_scale_scores()
head(ds_long)
dto[["social_network"]] <- ds_long

# ----- social-support -------------
#read in the renaming rules for this specific variables
rename_social_support  <-  ls_meta[["social-support"]] %>% as.data.frame()
# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  # year <- 2006
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_social_support,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

#for perceived social support(or relationship quality) all positive items need to be reverse coded.
rename_meta <-rename_social_support
reverse_these <- unique( rename_meta[rename_meta$reversed==TRUE,"new_name"] )
reverse_these <- reverse_these[!is.na(reverse_these)]
testit::assert("The scale does not contained reverse coded items",reverse_these==0L)
ds_long <- ds_long %>% 
  reverse_coding(reverse_these)

# Create a function that computes positive and negative social support totals for each category.
# compute_social_support_scale_score <- function(d){
positive_spouse <- c("ssup1sp","ssup2sp","ssup3sp")
ds_long[,"support_spouse_total"] <- apply(ds_long[positive_spouse],1,sum, na.rm = FALSE)
ds_long[,"support_spouse_mean"] <- apply(ds_long[positive_spouse],1,mean, na.rm = FALSE)

negative_spouse <- c("ssup4sp","ssup5sp","ssup6sp","ssup7sp")
ds_long[,"strain_spouse_total"] <- apply(ds_long[negative_spouse],1,sum, na.rm = FALSE)
ds_long[,"strain_spouse_mean"] <- apply(ds_long[negative_spouse],1,mean, na.rm = FALSE)

positive_child <- c("ssup1ch","ssup2ch","ssup3ch")
ds_long[,"support_child_total"] <- apply(ds_long[positive_child],1,sum, na.rm = FALSE)
ds_long[,"support_child_mean"] <- apply(ds_long[positive_child],1,mean, na.rm = FALSE)

negative_child <- c("ssup4ch","ssup5ch","ssup6ch","ssup7ch")
ds_long[,"strain_child_total"] <- apply(ds_long[negative_child],1,sum, na.rm = FALSE)
ds_long[,"strain_child_mean"] <- apply(ds_long[negative_child],1,mean, na.rm = FALSE)

positive_family <- c("ssup1fam","ssup2fam","ssup3fam")
ds_long[,"support_fam_total"] <- apply(ds_long[positive_family],1,sum, na.rm = FALSE)
ds_long[,"support_fam_mean"] <- apply(ds_long[positive_family],1,mean, na.rm = FALSE)

negative_family <- c("ssup4fam","ssup5fam","ssup6fam","ssup7fam")
ds_long[,"strain_family_total"] <- apply(ds_long[negative_family],1,sum, na.rm = FALSE)
ds_long[,"strain_family_total"] <- apply(ds_long[negative_family],1,mean, na.rm = FALSE)

positive_friends <- c("ssup1fr","ssup2fr","ssup3fr")
ds_long[,"support_friend_total"] <- apply(ds_long[positive_friends],1,sum, na.rm = FALSE)
ds_long[,"support_friend_mean"] <- apply(ds_long[positive_friends],1,mean, na.rm = FALSE)

negative_friends <- c("ssup4fr","ssup5fr","ssup6fr","ssup7fr")
ds_long[,"strain_friends_total"] <- apply(ds_long[negative_friends],1,sum, na.rm = FALSE)
ds_long[,"strain_friends_mean"] <- apply(ds_long[negative_friends],1,mean, na.rm = FALSE)
head(ds_long)

#   d$missing_count <- apply(d[col_names_7], 1, function(z) sum(is.na(z)))
#   d <- d %>% 
#     dplyr::mutate( 
#       score_wellbeing_7 = ifelse(missing_count<3, 
#                                  wellbeing_sum_7/(7- missing_count),NA))
#   d <- d %>% 
#     dplyr::mutate( 
#       wellbeing_sum_7 = ifelse(missing_count>0, NA, wellbeing_sum_7)
#     )
#   return(d)
# }
d <- ds_long %>% dplyr::filter(hhidpn==10001010)
#ds_long <- ds_long %>% compute_scale_score()
head(ds_long)
dto[["social_support"]] <- ds_long

# ---------- social-contact --------
#read in the renaming rules for this specific variable
rename_social_contact  <-  ls_meta[["social-contact"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()

for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_social_contact,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)

# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

# create a vector with names of items to be reverse scored 
# All social contact items are reversed scored 1 = Three or more times a week to 6 = less than once a year or never
rename_meta <-rename_social_contact
# use meta data to provide the rules
reverse_these <- unique( rename_meta[rename_meta$reversed==1,"new_name"] ) %>% as.data.frame()
reverse_these <- reverse_these[!is.na(reverse_these)]
ds_long <- ds_long %>% 
  reverse_coding(reverse_these)

#Create three social contact scores one for each group, taking the average of the three contact types
child_contact_vars <- c("mtchild","spkchild","wrtchild")
ds_long[,"children_contact_mean"] <- apply(ds_long[child_contact_vars],1,mean, na.rm = FALSE)

family_contact_vars <- c("mtfam","spkfam","wrtfam")
ds_long[,"family_contact_mean"] <- apply(ds_long[family_contact_vars],1,mean, na.rm = FALSE)

friend_contact_vars <- c("mtfriend","spkfriend","wrtfriend")
ds_long[,"friend_contact_mean"] <- apply(ds_long[friend_contact_vars],1,mean, na.rm = FALSE)

dto[["social_contact"]] <- ds_long

#--------activity--------
#read in the renaming rules for this specific variables
rename_activity   <-  ls_meta[["activity"]]
# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_activity,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)
# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_activity,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_activity,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

ds_long %>% dplyr::filter(hhidpn==10001010)

# create a vector with names of items to be reverse scored 
# All activity items are reversed scored 1 = Daily to 7 = Never/not relevant recode all so that higher numbers indicate more activity
rename_meta <-rename_activity
# use meta data to provide the rules
reverse_these <- unique( rename_meta[rename_meta$reversed==TRUE,"new_name"] ) %>% as.data.frame()
reverse_these <- reverse_these[!is.na(reverse_these)]
ds_long <- ds_long %>% 
  reverse_coding(reverse_these)

#Create activity summary scores
# d <- ds_long %>% dplyr::filter(hhidpn==10001010)

# take a subject
d_long <- ds_long %>% 
  dplyr::filter(year %in% c(2008, 2010, 2012, 2014)) %>%  # only these years
  dplyr::select(year, hhidpn, dplyr::matches("activity_")) # only these variables
# compute scale scores on subsetted ds
d_long <- compute_scale_score(d_long) %>% 
  dplyr::rename(activity_mean = mean, activity_sum = sum) %>% 
  dplyr::select(year, hhidpn, activity_mean, activity_sum)
# merge computed scales score to the general file
ds_long <- ds_long %>% dplyr::left_join(d_long)
dto[["activity"]] <- ds_long

#---------wellbeing------------
#read in the renaming rules for this specific variables
rename_wellbeing   <-  ls_meta[["wellbeing"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_wellbeing,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

ds_long %>% dplyr::filter(hhidpn==10001010)

# create a vector with names of items to be reverse scored 
rename_meta <-rename_wellbeing
# use meta data to provide the rules
reverse_these <- unique( rename_meta[rename_meta$reversed==TRUE,"new_name"] ) %>% as.data.frame()
reverse_these <- reverse_these[!is.na(reverse_these)]
ds_long <- ds_long %>% 
  reverse_coding(reverse_these)

compute_wellbeing_scale_score <- function(d){
  # d <- ds_long %>% dplyr::filter(hhidpn %in% c(3010,10281010))
  (col_names_7 <- setdiff(names(d),c("year","hhidpn")))
  (col_names_2 <- col_names_7[1:2])
  d[,"wellbeing_sum_7"] <- apply(d[col_names_7],1,sum, na.rm = TRUE)
  d[,"wellbeing_sum_2"] <- apply(d[col_names_2],1,sum, na.rm = TRUE)
  d[,"score_wellbeing_2"] <- apply(d[col_names_2],1,mean, na.rm = TRUE)
  d$missing_count <- apply(d[col_names_7], 1, function(z) sum(is.na(z)))
  d <- d %>% 
    dplyr::mutate( 
      score_wellbeing_7 = ifelse(missing_count<3, 
                                   wellbeing_sum_7/(7- missing_count),NA))
    d <- d %>% 
        dplyr::mutate( 
          wellbeing_sum_7 = ifelse(missing_count>0, NA, wellbeing_sum_7)
    )
  return(d)
}

ds_long <- ds_long %>% compute_wellbeing_scale_score()
head(ds_long)
# merge computed scales score to the general file

dto[["wellbeing"]] <- ds_long

#------self-rated-memory------
#read in the renaming rules for this specific variables
rename_self_rated_memory   <-  ls_meta[["self-rated-memory"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_self_rated_memory,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

#srmemory needs to be recoded and 8 means Don't know or Not Ascertained and 9 means Refused
ds_long[,"srmemory"] <- plyr::mapvalues(ds_long[,"srmemory"], from=c(9,8,5,4,3,2,1), to=c(NA,NA,1,2,3,4,5)) 

#srmemoryp is recoded such that higher numbers mean memory has improved and 8 and 9 are recoded as NA
ds_long[,"srmemoryp"] <- plyr::mapvalues(ds_long[,"srmemoryp"], from=c(9,8,3,2,1), to=c(NA,NA,1,2,3)) 

dto[["srmemory"]] <- ds_long

#---word-list-recall-------
rename_wordlist   <-  ls_meta[["word-list"]] 

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_wordlist,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

dto[["wordlist"]] <- ds_long

#---mental-status-----------
rename_mental_status   <-  ls_meta[["mental-status"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_mental_status,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)

mentalstatus_recode_vars <- c("msmonth", "msdate","msyear","msday","msnaming1","msnaming2","mspresident","msvp")
recoding_mentalstatus <- function(d, variables){
  for(v in variables){
    (p <- unique(d[,v]) %>% as.numeric())
    (p <- p[!is.na(p)])
    d[,v] <- plyr::mapvalues(d[,v], from=c(5,8,9), to=c(0,0,NA)) 
  }
  return(d)
}

# mental status items need to be recoded so that wrong answers are 0 (currently 5)
# and 8 means Don't know or Not Ascertained and is recoded as 0 (consistent with Ofstedal et al. 2005)
# 9 means Refused and is recoded as missing
ds_long <- ds_long %>% 
  recoding_mentalstatus(mentalstatus_recode_vars)
head(ds_long)
# recode count backwards variable so that 5 (incorrect) is 0 and 9 (refused or NA) is NA.
ds_long[,"countb"] <- plyr::mapvalues(ds_long[,"countb"], from = c(5,9), to =c(0,NA))
ds_long[,"countb2"] <- plyr::mapvalues(ds_long[,"countb2"], from = c(5,9), to =c(0,NA))
# if requested participants could start over if they requested this this should be used instead 
# create a count variable to be used
ds_long[,"count"] <- ifelse(ds_long[,"countb"]==6, ds_long[,"countb2"], ds_long[,"countb"])

mentalstatus_vars <- c("msmonth", "msdate","msyear","msday","msnaming1","msnaming2","mspresident","msvp","count")
# Calculate a mental status total score by summing the items
ds_long[,"mentalstatus_tot"] <- apply(ds_long[mentalstatus_vars],1,sum, na.rm = F)

tail(ds_long)

table(ds_long$mentalstatus_tot)

dto[["mentalstatus"]] <- ds_long

#-----vocabulary--------
rename_vocabulary   <-  ls_meta[["vocabulary"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_vocabulary,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

vocab_recode_vars <- c("vocab1", "vocab2","vocab3","vocab4","vocab5")
recoding_vocab<- function(d, variables){
  for(v in variables){
    (p <- unique(d[,v]) %>% as.numeric())
    (p <- p[!is.na(p)])
    d[,v] <- plyr::mapvalues(d[,v], from=9, to=NA) 
  }
  return(d)
}

# vocab items are coded with 9 as refused, recode this to NA scale is from 0 DK or incorrect to 2 perfectly correct.
ds_long <- ds_long %>% 
  recoding_vocab(vocab_recode_vars)

ds_long[,'vocab_total'] <-  apply(ds_long[vocab_recode_vars],1,sum, na.rm = FALSE)

dto[["vocabulary"]] <- ds_long

#-----depression--------------
rename_depression   <-  ls_meta[["depression"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_depression,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)
# ces-d items are coded with 9 as refused, 8 as dk, and 5 as no.
cesd_vars <- c("cesd1","cesd2","cesd3","cesd4","cesd5","cesd6","cesd7","cesd8")
recoding_depression<- function(d, variables){
  for(v in variables){
    (p <- unique(d[,v]) %>% as.numeric())
    (p <- p[!is.na(p)])
    d[,v] <- plyr::mapvalues(d[,v], from=c(5, 8, 9), to=c(0, NA, NA)) 
  }
  return(d)
}

ds_long <- recoding_depression(ds_long, cesd_vars)

cesd_vars_to_reverse <- c("cesd4","cesd6")
ds_long <- reverse_coding(ds_long, cesd_vars_to_reverse)

head(ds_long)

ds_long[,'dep_total'] <-  apply(ds_long[cesd_vars],1,sum, na.rm = FALSE)

dto[["depression"]] <- ds_long

#-------health--------------------
rename_health   <-  ls_meta[["health"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_health,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

healthconditions <- c("hypertension", "diabetes", "cancer", "lungdisease", "heart", "stroke", "arthritis") 

recoding_health<- function(d, variables){
  for(v in variables){
    (p <- unique(d[,v]) %>% as.numeric())
    (p <- p[!is.na(p)])
    d[,v] <- plyr::mapvalues(d[,v], from=c(3, 8, 9), to=c(0, NA, NA)) 
  }
  return(d)
}

ds_long[,"healthcond"] <- rowSums(ds_long[,healthconditions]==1)

exercise_vars <- c("vigorousactivity","moderateactivity","mildactivity")
recoding_exercise<- function(d, variables){
  for(v in variables){
    (p <- unique(d[,v]) %>% as.numeric())
    (p <- p[!is.na(p)])
    d[,v] <- plyr::mapvalues(d[,v], from=c(1, 2, 3, 4, 7, 8, 9), to=c(4, 3, 2, 1, 5, NA, NA)) 
  }
  return(d)
}
ds_long <- recoding_exercise(ds_long, exercise_vars)
ds_long[,"exercise"] <- apply(ds_long[exercise_vars],1,sum, na.rm = FALSE)

dto[["health"]] <- ds_long

# ------- serial-7's ---------
rename_serial   <-  ls_meta[["serial7s"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_serial,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

dto[["serial7s"]] <- ds_long

# ----- chronic-stressors -----------------------
rename_chronicstressors <- ls_meta[["chronic-stressors"]]

# now cycle through all ds for each year (must have ds_2004, ds_2006 objects)
ls_temp <- list()
for(year in c(2004, 2006, 2008, 2010, 2012, 2014)){ 
  # create a string to be passed as command to the eval() function
  cstring <- paste0(
    "ls_temp[[paste(year)]] <- subset_rename(ds_",year,", rename_chronicstressors,",year,")")
  eval(parse(text=cstring)) # evaluates the content of the command string
}
# this creates a list in which each element is a dataset
# each dataset contains items from target construct for that year
lapply(ls_temp,names)
# now we combine datasets from all years into a single LONG dataset
ds_long <- plyr::ldply(ls_temp, data.frame,.id = "year" ) %>% 
  dplyr::arrange(hhidpn)
head(ds_long)

chronicstress_vars <- c("healthprob","physicalprob","alcoholprob","workprob","financialprob","houseprob","relationshipprob","sickfam")

ds_long[,"stresstotal"] <- rowSums(ds_long[,chronicstress_vars])

dto[["chronic-stressors"]] <- ds_long



# # ---- save-to-disk ------------------------------------------------------------
# names(dto)
# lapply(dto, names)
# dto %>% object.size() %>% utils:::format.object_size("auto")
# # Save as a compress, binary R dataset.  It's no longer readable with a text editor, but it saves metadata (eg, factor information).
# saveRDS(dto, file="./data-unshared/derived/dto.rds", compress="xz")


# merge multiple datasets that are stored as elements of a list
merge_mulitple_files <- function(list, by_columns){
  Reduce(function( d_1, d_2 ) dplyr::full_join(d_1, d_2, by=by_columns), list)
}

#convert to factor for join
class(dto$rand$year)
head(dto$rand$year)
dto$rand$year <- as.factor(dto$rand$year)

ds_long <- merge_mulitple_files(dto, by_columns = c("year","hhidpn"))

# A list of the psychosocial variables to use to test for lb wave completion.
ds_lbvars <- ds_long %>% 
  dplyr::select(score_loneliness_3, score_loneliness_11, socialnetwork_total, snspouse, snchild, 
                snfamily, snfriends,support_spouse_total, support_child_total, support_fam_total, 
                support_friend_total, strain_spouse_total, strain_child_total, strain_family_total, 
                strain_friends_total, children_contact_mean, family_contact_mean, friend_contact_mean,
                activity_mean, activity_sum)

ds_long$lbvars <- rowSums(!is.na(ds_lbvars))
head(ds_long)

# ----- Creates a variable lbwave that designates wave number based on eligibility for the leave-behind questionnaire
for(i in unique(ds_long$hhidpn)){
  #id <- 3020
  id <- i
  dsyear <- ds_long %>% dplyr::filter(hhidpn==id)
  wavecount <- 0
  wave_04 <- 0
  
  for(y in unique(dsyear$year)){
    #year <- 2014
    year <- y
    current_row <- which(ds_long$hhidpn==id & ds_long$year==year)
    cond_2004 <- !is.na(ds_long[current_row,"lbgiven"]) & ds_long[current_row,"lbgiven"] == "QUESTIONNAIRE LEFT WITH RESPONDENT"
    if(cond_2004==TRUE){
      wave_04 <- wave_04+1
      ds_long[current_row,"lbwave"] <- wave_04
      next}
    wave_cond <- !is.na(ds_long[current_row,"lbeligibility"]) & ds_long[current_row,"lbeligibility"] == "Eligible for leave behind"
    if(wave_cond==TRUE){
      wavecount <- wavecount+1
      ds_long[current_row,"lbwave"] <- wave_04 + wavecount
    }else{
      ds_long[current_row,"lbwave"] <- 0 
    }
    wave_2014 <- ds_long[current_row,"lbvars"] > 0 & year == "2014"
    if(wave_2014==TRUE){
      wavecount2014 <- 1
      ds_long[current_row,"lbwave"] <- wave_04 + wavecount + wavecount2014
    }
  }}
ds_long %>% dplyr::filter(hhidpn== "3020")

if(ds_long$lbvars>0 & ds_long$lbwave==0)
  
  # ----- Creates a variable lbwave that designates wave number based on whether or not there are any
  # non-missing leave-behind questionnaire items. 
  for(i in unique(ds_long$hhidpn)){
    #id <- 3020
    id <- i
    dsyear <- ds_long %>% dplyr::filter(hhidpn==id)
    wavecount <- 0
    
    for(y in unique(dsyear$year)){
      #year <- 2012
      year <- y
      current_row <- which(ds_long$hhidpn==id & ds_long$year==year)
      
      wave_cond <- ds_long[current_row,"lbvars"] > 0
      if(wave_cond==TRUE){
        wavecount <- wavecount+1
        ds_long[current_row,"lbwave2"] <- wavecount
      }else{
        ds_long[current_row,"lbwave2"] <- 0 
      }
    }}


saveRDS(ds_long, file="./data-unshared/derived/data-long.rds")
# Examine distribution of lb waves
table(ds_long$lbwave2, ds_long$year)

(ds_long$lbwave2)

ds_long[which(ds_long$lbwave2==5),]

examine <- ds_long %>% dplyr::filter(hhidpn==500172010)


# select only cases where the participant belongs to one of the HRS cohorts.
# examine the descriptives over waves
over_waves <- function(ds, measure_name, exclude_values="") {
  ds <- as.data.frame(ds)
  testit::assert("No such measure in the dataset", measure_name %in% unique(names(ds)))
  # measure_name = "htval"; wave_name = "wave"; exclude_values = c(-99999, -1)
  cat("Measure : ", measure_name,"\n", sep="")
  t <- table( ds[,measure_name], ds[,"year"], useNA = "always"); t[t==0] <- ".";t
  print(t)
  cat("\n")
  ds[,measure_name] <- as.numeric(ds[,measure_name])
  
  d <- ds[!(ds[,measure_name] %in% exclude_values), ]
  a <- lazyeval::interp(~ round(mean(var),2) , var = as.name(measure_name))
  b <- lazyeval::interp(~ round(sd(var),3),   var = as.name(measure_name))
  c <- lazyeval::interp(~ n())
  dots <- list(a,b,c)
  t <- d %>%
    dplyr::select_("hhidpn","year", measure_name) %>%
    na.omit() %>%
    # dplyr::mutate_(measure_name = as.numeric(measure_name)) %>%
    dplyr::group_by_("year") %>%
    dplyr::summarize_(.dots = setNames(dots, c("mean","sd","count")))
  return(as.data.frame(t))
  
}

# ---- object-verification ------------------------------------------------
# the production of the dto object is now complete
# we verify its structure and content:
dto <- readRDS("./data-unshared/derived/dto.rds")
names(dto)
# at this point the dto contains elements
# each of which is a dataset with a subset of variables
# united by type of items
